Dagsats som long og skalert utbetalingsgrad

+++ domenetjenester/beregning-ytelse/src/main/java/no/nav/ung/sak/ytelse/beregning/TilkjentYtelsePeriode.java
@@ -6,7 +6,7 @@ import no.nav.ung.sak.domene.typer.tid.DatoIntervallEntitet;
 
 public record TilkjentYtelsePeriode(
     DatoIntervallEntitet periode,
-    BigDecimal dagsats,
+    Long dagsats,
     BigDecimal utbetalingsgrad
 ) {
 }
